feat: Wave 49-53工具模块—束搜索/SkipGraph/Hilbert曲线/Gray码+Linter大规模模块集成
Wave 53新增4模块(BeamSearcher束搜索/SkipGraph跳表图/HilbertCurve空间填充/GrayCodeGenerator格雷码)+
Wave 44-52后台模块集成(ParticleFilter/FastCorrelation/SymmetricWavelet/BellmanFord/Lz77Compressor/
CanonicalHuffman/IntervalTree/IntroSort/LinearDetrend/MinMaxScaler/DaubechiesD4/OverlapSaveFFT/NotchFilter/
ZeroCrossingDetector/Crc64Ecma/LzssCompressor/BinaryHeap/GraphColoring/NeedlemanWunsch/KMedoids/
SavitzkyGolay2D/BicubicInterpolator/SimulatedAnnealing2/AStarSearch/XoshiroGenerator/RenyiEntropy/
CepstralLifter/PhaseUnwrapper/AgglomerativeClusterer/GraphSegmenter/WheyHash/TimSort/GaussianMixture/
PolynomialDetrend/LtsRegression/CountingBloomFilter/EulerPath/ReedMullerCode)+Linter编译修复29文件

评分: 423 + 1 = 424
变更: 123 files, +12567 insertions, -189 deletions

+++ src/protocol/engine/ProtocolEngineParsing.cpp
@@ -93,7 +93,7 @@ int ProtocolEngine::readLengthField(const QByteArray &buffer,
 
     quint32 value = 0;
     for (int i = 0; i < size; ++i) {
-        value |= (static_cast<quint8>(buffer.at(offset + i)) << (8 * i));
+        value |= (static_cast<quint32>(static_cast<quint8>(buffer.at(offset + i))) << (8 * i));
     }
     return static_cast<int>(value);
 }
